Fix Debian-specific checking for dmenu.
The old check uses ``which``, which is a Debian-specific tool.  The new one
invokes ``dmenu -v`` straight away, so it doesn't depend on external tools.
It's compatible with both Python 3 and Python 2.

It also fixes opening ``/dev/null`` in write mode instead of read-only, plus
it uses ``with`` to ensure that the file is closed.